refactor: Member 도메인 Enum 분리
