+++ src/main/java/com/dongyang/dongpo/domain/member/enums/Gender.java
@@ -1,0 +1,5 @@
+package com.dongyang.dongpo.domain.member.enums;
+
+public enum Gender {
+    GEN_MALE, NONE, GEN_FEMALE
+}
