+++ src/main/java/com/dongyang/dongpo/domain/member/enums/Role.java
@@ -1,0 +1,5 @@
+package com.dongyang.dongpo.domain.member.enums;
+
+public enum Role {
+    ROLE_MEMBER, ROLE_ADMIN
+}
