+++ src/main/java/com/dongyang/dongpo/domain/member/enums/SocialType.java
@@ -1,0 +1,5 @@
+package com.dongyang.dongpo.domain.member.enums;
+
+public enum SocialType {
+    KAKAO, APPLE, NAVER
+}
