+++ src/main/java/com/dongyang/dongpo/domain/member/enums/Status.java
@@ -1,0 +1,5 @@
+package com.dongyang.dongpo.domain.member.enums;
+
+public enum Status {
+    ACTIVE, INACTIVE, LEAVE
+}
