+++ src/main/java/com/dongyang/dongpo/domain/member/repository/MemberTitleRepository.java
@@ -2,7 +2,7 @@ package com.dongyang.dongpo.domain.member.repository;
 
 import com.dongyang.dongpo.domain.member.entity.Member;
 import com.dongyang.dongpo.domain.member.entity.MemberTitle;
-import com.dongyang.dongpo.domain.member.entity.Title;
+import com.dongyang.dongpo.domain.member.enums.Title;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
